backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
anthropic==0.39.0
openai==1.54.3
python-multipart==0.0.12
gunicorn == 23.0.0